fix: exclude test utils from transpilation and rename to remove .test
