+++ tsconfig.json
@@ -6,5 +6,5 @@
     "outDir": "lib"
   },
   "include": ["src/**/*"],
-  "exclude": ["src/**/*.test.ts"]
+  "exclude": ["src/**/*.test.ts", "src/testUtils.ts"]
 }
